#144 [rename] adapter->persistence->repository 내부에 추후 쿼리dsl 도입을 위하여 query, jpa 패키지로 패키지 구조를 변경한다.
