+++ src/main/java/com/dataracy/modules/data/adapter/persistence/impl/DataMetadataRepositoryAdapter.java
@@ -3,8 +3,8 @@ package com.dataracy.modules.data.adapter.persistence.impl;
 import com.dataracy.modules.data.adapter.persistence.entity.DataEntity;
 import com.dataracy.modules.data.adapter.persistence.entity.DataMetadataEntity;
 import com.dataracy.modules.data.adapter.persistence.mapper.DataMetadataEntityMapper;
-import com.dataracy.modules.data.adapter.persistence.repository.DataJpaRepository;
-import com.dataracy.modules.data.adapter.persistence.repository.DataMetadataJpaRepository;
+import com.dataracy.modules.data.adapter.persistence.repository.jpa.DataJpaRepository;
+import com.dataracy.modules.data.adapter.persistence.repository.jpa.DataMetadataJpaRepository;
 import com.dataracy.modules.data.application.port.out.DataMetadataRepositoryPort;
 import com.dataracy.modules.data.domain.exception.DataException;
 import com.dataracy.modules.data.domain.model.DataMetadata;
