+++ src/main/java/com/dataracy/modules/data/adapter/persistence/repository/jpa/DataMetadataJpaRepository.java
@@ -1,4 +1,4 @@
-package com.dataracy.modules.data.adapter.persistence.repository;
+package com.dataracy.modules.data.adapter.persistence.repository.jpa;
 
 import com.dataracy.modules.data.adapter.persistence.entity.DataMetadataEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
