+++ src/main/java/com/dataracy/modules/reference/adapter/persistence/repository/jpa/DataSourceJpaRepository.java
@@ -1,4 +1,4 @@
-package com.dataracy.modules.reference.adapter.persistence.repository;
+package com.dataracy.modules.reference.adapter.persistence.repository.jpa;
 
 import com.dataracy.modules.reference.adapter.persistence.entity.DataSourceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
