+++ src/main/java/com/dataracy/modules/reference/adapter/persistence/repository/jpa/OccupationJpaRepository.java
@@ -1,4 +1,4 @@
-package com.dataracy.modules.reference.adapter.persistence.repository;
+package com.dataracy.modules.reference.adapter.persistence.repository.jpa;
 
 import com.dataracy.modules.reference.adapter.persistence.entity.OccupationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
